Separate the fancy implode() logic out of breema_get_entity_label_multiple():
We want to use this code for things that aren't only arrays of entity
objects (e.g. both Breema and non-Breema authors on articles), so put
all the fancy implode() stuff dealing with the right delimiters into a
separate function and keep the entity callback stuff out of the way.