Share: Fix bug in meta tag generation code while cloning entities:
While cloning a node, we can have a node parameter that's not a
fully-loaded entity, leading to fatal errors in our meta tag
generation code. We don't care about meta tags at this point, but we
don't want to crash, either. So make sure the node we have is an
instance of NodeInterface (a fully loaded entity) and not a string
while generating the meta tags.

This bug was introduced in commit 0b780eba.