Menu: Consistent padding on horizontal menu links for vertical alignment:
The default is 0.75em padding (smallest screens w/ horizontal menu).
When we do the media query to bump that to 1em, we also need to bump
the padding on the sub-menu links so everything still lines up.
Also bump the width of a few sub-menus to ensure words don't wrap.