User: Change save button on event browser to 'Save selection':
The button to open the modal is already called 'Select events'.
The modal's title is also 'Select events'.
The action inside the modal is to save what you already selected.
Update the help text on 'Attended classes' to match.
Also warn the users to click 'Save selection' on each page.